requirements.txt
django-bootstrap4
# django-datatable
django-tables2
